I'm working out the back and forth of information between the add course screen and to add an assessment. The info is being passed ok, but the assessment isn't saving. I'm thinking it is something to do with the get context on line 292 (being in the activity for Courses and not Assessments). It's late, I'll sleep on it.
I corrected the constraints on the instructor activity. In the course adapter I started created a copy of the Course List, copyCourses (line 26), to use for a filter later. I did the same in the course details - added Lists for notes and assessments (line 153 and 154) to use for filters later.

I did some minor changes in the course details fragment_course_details.xml for constraints.